refactor: move SharedScreenComponents to components package and update related imports
