+++ app/src/main/java/at/aau/serg/scotlandyard/ui/components/SharedScreenComponents.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.scotlandyard.ui.activity
+package at.aau.serg.scotlandyard.ui.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -10,7 +10,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
